+++ src/uni_modules/wot-design-uni/components/wd-calendar-view/year/year.vue
@@ -28,6 +28,7 @@ export default {
 </script>
 
 <script lang="ts" setup>
+import wdToast from '../../wd-toast/wd-toast.vue'
 import { computed, ref, watch } from 'vue'
 import { deepClone, isArray, isFunction } from '../../common/util'
 import { compareMonth, formatYearTitle, getDateByDefaultTime, getItemClass, getMonthByOffset, getMonthOffset } from '../utils'