+++ src/uni_modules/wot-design-uni/components/wd-form/wd-form.vue
@@ -17,13 +17,14 @@ export default {
 </script>
 
 <script lang="ts" setup>
+import wdToast from '../wd-toast/wd-toast.vue'
 import { reactive, watch } from 'vue'
 import { deepClone, getPropByPath, isDef, isPromise } from '../common/util'
 import { useChildren } from '../composables/useChildren'
 import { useToast } from '../wd-toast'
 import { type FormRules, FORM_KEY, type ErrorMessage, formProps, type FormExpose } from './types'
 
-const toast = useToast('wd-form-toast')
+const { show: showToast } = useToast('wd-form-toast')
 const props = defineProps(formProps)
 
 const { children, linkChildren } = useChildren(FORM_KEY)
@@ -117,9 +118,7 @@ async function validate(prop?: string): Promise<{ valid: boolean; errors: ErrorM
 
   await Promise.all(promises)
 
-  errors.forEach((error) => {
-    showMessage(error)
-  })
+  showMessage(errors)
 
   if (valid) {
     if (prop) {
@@ -150,13 +149,16 @@ function getMergeRules() {
   return mergedRules
 }
 
-function showMessage(errorMsg: ErrorMessage) {
-  if (!errorMsg.message) return
-
-  if (props.errorType === 'toast') {
-    toast.show(errorMsg.message)
-  } else if (props.errorType === 'message') {
-    errorMessages[errorMsg.prop] = errorMsg.message
+function showMessage(errors: ErrorMessage[]) {
+  const messages = errors.filter((error) => error.message)
+  if (messages.length) {
+    if (props.errorType === 'toast') {
+      showToast(messages[0].message)
+    } else if (props.errorType === 'message') {
+      messages.forEach((error) => {
+        errorMessages[error.prop] = error.message
+      })
+    }
   }
 }