+++ src/uni_modules/wot-design-uni/components/wd-index-bar/type.ts
@@ -1,14 +1,8 @@
-import type { Ref } from 'vue'
 import type { InjectionKey } from 'vue'
 import type { ExtractPropTypes } from 'vue'
 import { makeBooleanProp } from '../common/props'
 
 export type AnchorIndex = number | string
-
-export interface AnchorItem {
-  top: number
-  index: AnchorIndex
-}
 
 export const indexBarProps = {
   /**
@@ -22,7 +16,6 @@ export type IndexBarProps = ExtractPropTypes<typeof indexBarProps>
 export type InderBarProvide = {
   props: { sticky?: boolean }
   anchorState: {
-    anchorList: AnchorItem[] // 锚点列表
     activeIndex: AnchorIndex | null // 当前激活的索引
   }
 }