+++ src/uni_modules/wot-design-uni/components/wd-picker-view/wd-picker-view.vue
@@ -43,6 +43,7 @@ export default {
 }
 </script>
 <script lang="ts" setup>
+import wdLoading from '../wd-loading/wd-loading.vue'
 import { getCurrentInstance, ref, watch, nextTick } from 'vue'
 import { deepClone, getType, isArray, isDef, isEqual, range } from '../common/util'
 import { formatArray, pickerViewProps, type ColumnItem, type PickerViewExpose } from './types'
@@ -326,6 +327,15 @@ function getSelectedIndex() {
   return selectedIndex.value
 }
 
+/**
+ * 用于重置列数据为指定列数据
+ */
+function resetColumns(columns: (string | number | string[] | number[] | ColumnItem | ColumnItem[])[]) {
+  if (isArray(columns) && columns.length) {
+    formatColumns.value = formatArray(columns, props.valueKey, props.labelKey)
+  }
+}
+
 function onPickStart() {
   emit('pickstart')
 }
@@ -342,7 +352,8 @@ defineExpose<PickerViewExpose>({
   getColumnData,
   getColumnIndex,
   getLabels,
-  getSelectedIndex
+  getSelectedIndex,
+  resetColumns
 })
 </script>
 <style lang="scss" scoped>