+++ src/uni_modules/wot-design-uni/components/wd-progress/wd-progress.vue
@@ -27,6 +27,7 @@ export default {
 </script>
 
 <script lang="ts" setup>
+import wdIcon from '../wd-icon/wd-icon.vue'
 import { computed, ref, watch } from 'vue'
 import { checkNumRange, isArray, objToStyle } from '../common/util'
 import { progressProps } from './types'