+++ src/uni_modules/wot-design-uni/components/wd-select-picker/wd-select-picker.vue
@@ -115,6 +115,14 @@ export default {
 </script>
 
 <script lang="ts" setup>
+import wdActionSheet from '../wd-action-sheet/wd-action-sheet.vue'
+import wdCheckbox from '../wd-checkbox/wd-checkbox.vue'
+import wdCheckboxGroup from '../wd-checkbox-group/wd-checkbox-group.vue'
+import wdRadio from '../wd-radio/wd-radio.vue'
+import wdRadioGroup from '../wd-radio-group/wd-radio-group.vue'
+import wdButton from '../wd-button/wd-button.vue'
+import wdLoading from '../wd-loading/wd-loading.vue'
+
 import { getCurrentInstance, onBeforeMount, ref, watch, nextTick, computed } from 'vue'
 import { useCell } from '../composables/useCell'
 import { getRect, isArray, isDef, isFunction, requestAnimationFrame } from '../common/util'