+++ src/uni_modules/wot-design-uni/components/wd-upload/wd-upload.vue
@@ -94,6 +94,10 @@ export default {
 </script>
 
 <script lang="ts" setup>
+import wdIcon from '../wd-icon/wd-icon.vue'
+import wdVideoPreview from '../wd-video-preview/wd-video-preview.vue'
+import wdLoading from '../wd-loading/wd-loading.vue'
+
 import { computed, ref, watch } from 'vue'
 import { context, getType, isEqual, isImageUrl, isVideoUrl, isFunction, isDef } from '../common/util'
 import { chooseFile } from './utils'